backend/requirements.txt
flask
fastapi
uvicorn
supabase
requests
python-dotenv
crewai
